+++ openfl-workspace/tf_3dunet_brats/requirements.txt
@@ -1,6 +1,6 @@
+keras==3.6.0
 nibabel
 numpy
 
 setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
-keras==3.6.0
 tensorflow==2.18.0
