+++ openfl-workspace/tf_3dunet_brats/src/define_model.py
@@ -4,6 +4,7 @@
 """You may copy this file as the starting point of your own model."""
 
 import tensorflow as tf
+import keras
 
 
 def dice_coef(target, prediction, axis=(1, 2, 3), smooth=0.0001):
@@ -97,13 +98,13 @@ def build_model(input_shape,
     if (input_shape[0] % (2**depth)) > 0:
         raise ValueError(f'Crop dimension must be a multiple of 2^(depth of U-Net) = {2**depth}')
 
-    inputs = tf.keras.layers.Input(input_shape, name='brats_mr_image')
+    inputs = keras.layers.Input(input_shape, name='brats_mr_image')
 
-    activation = tf.keras.activations.relu
+    activation = keras.activations.relu
 
     params = {'kernel_size': (3, 3, 3), 'activation': activation,
               'padding': 'same',
-              'kernel_initializer': tf.keras.initializers.he_uniform(seed=seed)}
+              'kernel_initializer': keras.initializers.he_uniform(seed=seed)}
 
     convb_layers = {}
 
@@ -111,48 +112,48 @@ def build_model(input_shape,
     filters = initial_filters
     for i in range(depth):
         name = f'conv{i + 1}a'
-        net = tf.keras.layers.Conv3D(name=name, filters=filters, **params)(net)
+        net = keras.layers.Conv3D(name=name, filters=filters, **params)(net)
         if i in dropout_at:
-            net = tf.keras.layers.Dropout(dropout)(net)
+            net = keras.layers.Dropout(dropout)(net)
         name = f'conv{i + 1}b'
-        net = tf.keras.layers.Conv3D(name=name, filters=filters, **params)(net)
+        net = keras.layers.Conv3D(name=name, filters=filters, **params)(net)
         if batch_norm:
-            net = tf.keras.layers.BatchNormalization()(net)
+            net = keras.layers.BatchNormalization()(net)
         convb_layers[name] = net
         # only pool if not last level
         if i != depth - 1:
             name = f'pool{i + 1}'
-            net = tf.keras.layers.MaxPooling3D(name=name, pool_size=(2, 2, 2))(net)
+            net = keras.layers.MaxPooling3D(name=name, pool_size=(2, 2, 2))(net)
             filters *= 2
 
     # do the up levels
     filters //= 2
     for i in range(depth - 1):
         if use_upsampling:
-            up = tf.keras.layers.UpSampling3D(
+            up = keras.layers.UpSampling3D(
                 name=f'up{depth + i + 1}', size=(2, 2, 2))(net)
         else:
-            up = tf.keras.layers.Conv3DTranspose(name=f'transConv{depth + i + 1}',
+            up = keras.layers.Conv3DTranspose(name=f'transConv{depth + i + 1}',
                                                  filters=filters,
                                                  kernel_size=(2, 2, 2),
                                                  strides=(2, 2, 2),
                                                  padding='same')(net)
-        net = tf.keras.layers.concatenate(
+        net = keras.layers.concatenate(
             [up, convb_layers[f'conv{depth - i - 1}b']],
             axis=-1
         )
-        net = tf.keras.layers.Conv3D(
+        net = keras.layers.Conv3D(
             name=f'conv{depth + i + 1}a',
             filters=filters, **params)(net)
-        net = tf.keras.layers.Conv3D(
+        net = keras.layers.Conv3D(
             name=f'conv{depth + i + 1}b',
             filters=filters, **params)(net)
         filters //= 2
 
-    net = tf.keras.layers.Conv3D(name='prediction', filters=n_cl_out,
+    net = keras.layers.Conv3D(name='prediction', filters=n_cl_out,
                                  kernel_size=(1, 1, 1),
                                  activation='sigmoid')(net)
 
-    model = tf.keras.models.Model(inputs=[inputs], outputs=[net])
+    model = keras.models.Model(inputs=[inputs], outputs=[net])
 
     return model
