+++ openfl-workspace/tf_cnn_histology/requirements.txt
@@ -1,4 +1,4 @@
-pillow
 keras==3.6.0
+pillow
 tensorflow==2.18.0
 tensorflow-datasets
